Rewrote train loops

+++ src/models/GraphALS/utils.py
@@ -2,7 +2,7 @@ import torch
 
 
 def left_normalize_adj(adj:torch.sparse.Tensor):
-    degrees = adj.to_dense().sum(dim=1)
+    degrees = adj.to_dense().sum(dim=1).float()
     degrees = 1/degrees
     print(f'normalization matrix {degrees.size()}')
     print(f'adjacency matrix {adj.size()}')
